[flint] MFT 4.23.0-79 GUID range issue - support for MNVDA query
Description: fixed MACs/GUIDs query on device to take into cosideration the msb byte that was added to the guids struct, in order to support larger number of guids for albatross.

Tested OS: linux
Tested devices: image
Tested flows: flint sg/smg

Known gaps (with RM ticket): N/A

Issue: 3390705

+++ fw_comps_mgr/fw_comps_mgr.h
@@ -91,7 +91,7 @@ typedef f_prog_func_adv_st ProgressCallBackAdvSt;
 
 struct uid_entry
 {
-    u_int8_t num_allocated;
+    u_int16_t num_allocated;
     u_int64_t uid;
 };
